refactor(memory): Clean Code cleanup + store perf optimizations
Clean Code pass (P0-P3 from review):

- remove dead code: EntityNormalizer (unused; leaked rusqlite into entity service), EntityNotFound variant, LongInactivity evidence type, empty embedding_store.rs placeholder

- dedupe: store parse fns reuse FromStr (added FromStr to ConceptType/ObservationSourceType/SourceType); ingest flush logic -> flush_section; observation insert/insert_batch share obs_insert_sql

- correctness: parse_concept_type returns Option (was wrongly defaulting unknown -> Architecture); extract parse_source_type now warns; rewrite no-op all_evidence_types_have_weights test into a real coverage assertion; remove misleading CURRENT_VERSION in migration

- clippy: strip_suffix, drop redundant +Sync, needless_question_mark; CLI main returns Result; BetaConfidence f32->f64; concept full-field round-trip test locks column-positional binding

Perf:

- raw_memory insert_batch (single transaction) wired into CLI ingest, replacing N per-row lock acquisitions

- LazyLock-cached SQL strings (concept/observation stores); list queries consolidated via (?2 IS NULL OR status = ?2)

- unify collect_candidates/collect_concepts into generic map_rows

Verified: cargo check 0 warnings, clippy clean, 39/0 tests, CLI ingest smoke.

+++ crates/memory-runtime/tests/integration_test.rs
@@ -112,7 +112,7 @@ fn test_full_pipeline_json_ingest_to_store() {
 #[test]
 fn test_confidence_integration() {
     let mut bc = BetaConfidence::new();
-    assert!((bc.confidence() - 0.5).abs() < f32::EPSILON);
+    assert!((bc.confidence() - 0.5).abs() < f64::EPSILON);
 
     bc.update(&EvidenceType::UserConfirmation);
     assert!(bc.confidence() > 0.5);
